+++ app/src/main/java/com/nervousfish/nervousfish/exceptions/DatabaseException.java
@@ -1,4 +1,4 @@
-package com.nervousfish.nervousfish.modules.database;
+package com.nervousfish.nervousfish.exceptions;
 
 import java.io.IOException;
 import java.io.ObjectInputStream;
