+++ app/src/test/java/com/nervousfish/nervousfish/ConstantKeywordsTest.java
@@ -6,9 +6,11 @@ import static com.nervousfish.nervousfish.BaseTest.accessConstructor;
 import static org.junit.Assert.assertNotNull;
 
 public final class ConstantKeywordsTest {
+
     @Test
     public void testConstantKeywords() {
         ConstantKeywords keywords = (ConstantKeywords) accessConstructor(ConstantKeywords.class);
         assertNotNull(keywords);
     }
+
 }
