+++ app/src/test/java/com/nervousfish/nervousfish/exceptions/NoBluetoothExceptionTest.java
@@ -2,7 +2,6 @@ package com.nervousfish.nervousfish.exceptions;
 
 
 import com.nervousfish.nervousfish.TestException;
-import com.nervousfish.nervousfish.modules.database.DatabaseException;
 
 import org.junit.Test;
